feat : 사다리 구성 클래스 수정
- 사다리 요소를 관리하는 자료구조를 리스트로 변경하였습니다.
- 사다리 간격이 5로 변경되었으므로 출력 기호를 수정하였습니다.
