--- src/main/model/Game.java
@@ -1,20 +1,0 @@
-package main.model;
-
-public class Game {
-
-    private int numOfPeople;
-    private int maxHeightOfLadder;
-
-    private Ladder ladder;
-
-    public void init(int numOfPeople, int maxHeightOfLadder) {
-        this.numOfPeople = numOfPeople;
-        this.maxHeightOfLadder = maxHeightOfLadder;
-        ladder = new Ladder(numOfPeople, maxHeightOfLadder);
-    }
-
-    public String[][] getLadderBoard() {
-        return ladder.getBoard();
-    }
-
-}
